fix labels so they have the correct offsets
GNU-R bytecode contains both instructions and metadata, ours only contains instructions, so the indices of our instructions are different. Since labels are offsets in the bytecode, this means they need to be different as well.

Of course, if the labels are only used by GNU-R and we don't modify and re-encode the bytecode this is unnecessary. But, I'm sure we'll use the labels and their offsets while compiling, even if just to find the position to generate native code