move snapshots into `src/test` and combine `src/test/R` and `src/test/resources`
Move snapshots into `src/test` so they're closer to the resources they get generated from and are easier to find.

Combine `src/test/R` and `src/test/resources` because the distinction is confusing and adds extra complexity we don't really need.

Also add a few comments to clarify how the resources and snapshots are used.