add `ir.type` and start CFG.
@WIP init package, CFG, BB, Node, Value, ValueType, Instr

@WIP refine

@WIP add instructions list

initial value-type writeup

@WIP CFG, instrs, ValueType

@WIP draft RType implementation

@WIP drafting Stmts, Jumps, ...

@WIP more work...

@WIP minor fixes to the types and replace function

@WIP more BB and CFG work

@WIP try putting inheritance in types

it's cleaner to the user (less parallel representation), but a lot urliger to implement...I probably need to see if there's a better way to implement without as much duplication, or if the lack of parallel representation is worth it

@WIP new `RType` design

`RType` is a union of `RSexpType`s, which are for particular types and special values (currently functions, primitive vectors, the missing value, and everything else)

@WIP redo the type system

@WIP make it compile + remove most GNU-R bytecodes

@WIP type system improvements

@WIP arbitrary providers + bugfixes

TODO add providers for function, primitive vector, and generic value types which aren't exact, and fix the tests

@WIP bugfixes

@WIP bugfixes

main issues is that jqwik gives stack overflows trying to shrink the generated results. I still need to test non-trivial cases and ideally would like shrinking, but if I can't figure out why it doesn't work I'll have to disable it.

@WIP fixed jqwik generation, the issue now seems to be with function types

@WIP

@WIP dominator tree

@WIP simplify RType arbitrary

@WIP update notes

@WIP redo function types much better (untested)

includes re-implementing `Rf_matchArgs_NR` (`match.args`?)

@WIP add `desc`s to instructions and fix to build

comment code + test fixes to get this to build

@WIP fix function types and tests

@WIP really fix function types and tests

Property tests run in OK time (<1min) and haven't gotten a failure yet.

@WIP document `RType` in notes

How it works and explains some decisions.

It may change a lot from here though.

@WIP various fixups

@WIP simplify RType based on feedback

- `RType` is no longer a union, just has one `RValueType`. `RValueType`'s name is more accurate.
- The missing type is better represented; it's orthogonal like `RPromiseType`, except that `isMissing = YES` implies `value = null`. This way, we can still represent "known type OR missing".
- No more numeric primitive vector. I kept "numeric or logical" in case because binary operators support all of them, but not the non-numeric string and raw (comparison operators don't care).
- Potential bugfixes