resolve VSCode warnings (again)
There are still a couple warnings:

- A couple null checks which are immediately apparent but VSCode can't decide. If this is an issue, we should disable null analysis entirely on VSCode.
- "The hierarchy of type X is inconsistent" but it compiles fine. This seems to be an Eclipse/VSCode issue and I think the only way to resolve is to simplify inheritance, but idk if it's worth it since this only shows up in analysis.

Also, VSCode seems to change the non-null settings on startup which caused the warnings to pop-up again before. I also changed VSCode settings, so if this happens, it will simply disable all null-checks instead of breaking them. So, maybe we'll just disable null analysis entirely on VSCode...