finish draft closure compiler
- (frame states and promises)
+ initial tests

@WIP draft implementation of bc->PIR-IR

some things are probably not correct though, also a lot of specialized GNU-R bytecodes get converted into CallBuiltin because we don't have PIR instructions with the same specialization

some things are also still not implemented (e.g. `MakeClosure`)

@WIP fix bugs and add documentation to the draft implementation

@WIP closure and closure version overhaul and begin their compiler

also brainstorm high-level (how compilation/evaluation will work)

@WIP closure and closure version compiler + created `Module`

WIP:

- How promises will be compiled (and where they're needed).
- Add new PIR instructions to implement missing functionality required to compile some CFG bytecodes.
- (Maybe longer term) try to implement PushContext and PopContext because unless I'm mistaken they are created by GNU-R wherever there's a `next` or `break`, whereas RIR only needs them for niche complex cases.

@WIP improve `CFGTests`

Report test differently if we failed to parse, but in an acceptable way, so it will be apparent (unfortunately only to the human) if a large amount of tests are failing this way (which is not acceptable).

@WIP further improve `CFGPirTests`

and further fix `CFGEdit`

@WIP improve `CFGCompiler`

- create a call stack instead of putting the function on the regular stack and having a call arguments stack (this may have fixed semantics)

cleanup `Compiler` warnings + other small refactors

@WIP finish draft closure compiler (frame-states and promises) + initial tests