@WIP IR: it builds...
Some of the tests for non-IR parts of the compiler don't pass, but I have to investigate, because it's the compiler inlining things that GNU-R doesn't.

The IR is very unfinished, I had to disable the tests becase they wouldn't complete. `LatticeTests` also don't work, though they fail fast on relations that aren't yet fully implemented.