Add automated benchmarking (#1)
Add benchmark workflow to the CI.

It consists of two jobs:
1. Images - builds the necessary docker images
2. Benchmarks - runs the benchmark suite we use and package the results as downloadable artifacts

The first job runs in the GitHub cloud. The benchmarking job runs sequentially on a machine in the lab, i.e., in somewhat controlled environment.

## Images

There are 3 images:

- `rcp-base` - the base image which should have all the system packages to build the code and run benchmarks
- `rcp-rsh` - contains the modified version of the R virtual machine
- `rcp` - contains this project

## Benchmarks

Using the `rcp` it runs two benchmarks. The baseline which uses R byte code interpreter and one which uses the R the copy and patch JIT.