don't run checks after `Cleanup`
which substitutes origins and temporarily produces an invalid CFG, but it always gets fixed by other optimizations later.

This causes FIŘ tests to pass, but it's a stopgap. We may need to disable for a few passes after `Cleanup` (until the pass that inlines promises) or come up with a better solution