@WIP fix interpreter trace testing
specifically the traces' deoptimization states are equal at checkpoints. Not their entire state.

It catches one of the errors. I don't know if it significantly slows or hangs one of the later tests though, because I'm testing on my macbook which is slower; it probably significantly slows but doesn't hang, because there's lots of deep copies to get snapshots.